requirements.txt
flask
python-dotenv
pinecone
langchain
langchain-core
langchain-community
langchain-pinecone
langchain-cohere
cohere
pypdf
requests
urllib3